backend/agent_flow/requirements.txt
google-adk
google-generativeai
google-genai
python-dotenv
jinja2
perspective.py
qdrant-client
sentence-transformers
torch
